Move the window surfaces, the palette shell and the composition root
Windows/ takes the dialog, HUD, aux-window and About code — including
AuxWindowController, which had been living at the bottom of AboutView.
Palette/ takes the panel, controller, root view, screen protocol and the
palette state types extracted from AppCore. Contents unchanged.
