+++ Tinycast/Palette/PaletteCoordinator.swift
