+++ Tinycast/Palette/PaletteMode.swift
@@ -1,0 +1,63 @@
+import AppKit
+
+enum PaletteMode: String, CaseIterable, Identifiable {
+    case launcher
+    case clipboard
+    case calculatorHistory
+    case emoji
+    case uninstall
+    case quicklinks
+    /// Collects a quicklink's `{argument}` values before it opens; the pending request lives on
+    /// `AppCore.quicklinkArguments`, the way `.uninstall`'s target lives on `UninstallSession`.
+    case quicklinkArguments
+
+    var id: String { rawValue }
+    var title: String {
+        switch self {
+        case .launcher: return "Apps"
+        case .clipboard: return "Clipboard"
+        case .calculatorHistory: return "Calculator History"
+        case .emoji: return "Emoji & Symbols"
+        case .uninstall: return "Uninstall Application"
+        case .quicklinks: return "Quicklinks"
+        case .quicklinkArguments: return "Open Quicklink"
+        }
+    }
+    var systemImage: String {
+        switch self {
+        case .launcher: return "magnifyingglass"
+        case .clipboard: return "doc.on.doc"
+        case .calculatorHistory: return "plus.forwardslash.minus"
+        case .emoji: return "face.smiling"
+        case .uninstall: return "trash"
+        case .quicklinks, .quicklinkArguments: return Quicklink.sfSymbol
+        }
+    }
+    var placeholder: String {
+        switch self {
+        case .launcher: return "Search for apps and commands…"
+        case .clipboard: return "Type to filter entries…"
+        case .calculatorHistory: return "Do math, convert units, or search your past calculations…"
+        case .emoji: return "Search emoji and symbols…"
+        case .uninstall: return "Filter files and folders by name…"
+        case .quicklinks: return "Search quicklinks…"
+        // Replaced by the pending argument's name; only reached if the session vanished mid-render.
+        case .quicklinkArguments: return "Enter a value…"
+        }
+    }
+}
+
+/// The app a paste will land in, resolved once per palette show so the footer pill and menu rows can name it without re-reading `NSWorkspace` on every render.
+struct PasteTarget: Equatable {
+    let name: String
+    /// Bundle path for `IconCache` — nil for a target with no on-disk bundle.
+    let iconPath: String?
+
+    init?(app: NSRunningApplication?) {
+        guard let app, let name = app.localizedName else { return nil }
+        self.name = name
+        iconPath = app.bundleURL?.path
+    }
+
+    var pasteTitle: String { "Paste to \(name)" }
+}
