+++ Tinycast/Palette/PalettePanel.swift
