+++ Tinycast/Palette/PaletteScreen.swift
