+++ Tinycast/Palette/PaletteViewModel.swift
@@ -1,0 +1,37 @@
+import Foundation
+
+/// View-model shared between the panel's SwiftUI tree and the coordinator.
+@MainActor
+@Observable
+final class PaletteViewModel {
+    var mode: PaletteMode = .launcher
+    var query: String = ""
+    var selection: Int = 0
+    /// Changes every time the palette is shown so the search field can re-focus.
+    var focusToken = UUID()
+    /// Changes only when `prepare` resets the palette, so the lists snap their scroll to the top even when query/mode were already at their defaults (`focusToken` can't serve: it bumps on every reopen, which must preserve a within-timeout scroll).
+    var resetToken = UUID()
+    /// Changes when an action reorders the list under the selection (pinning a clip lifts it into the Pinned section), so the list scrolls the highlight back into view.
+    var followToken = UUID()
+    /// Set by the compact bar's "…" overflow to expand into the full launcher without a query; cleared on every `prepare`.
+    var forceExpanded = false
+    /// The app a paste would land in, mirrored from `PaletteWindowController.previousApp` on every show. Deliberately *not* cleared by `prepare` — pop-to-root resets the screen, not the paste target.
+    var pasteTarget: PasteTarget?
+    /// Gates the mouse-hover highlight: true only while the pointer is physically moving (armed on `.mouseMoved`, disarmed on any `.keyDown` in `PalettePanel.sendEvent`). Untracked — read at hover time, never drives a re-render.
+    @ObservationIgnored var hoverHighlightArmed = false
+    /// True while a footer popover menu (⌘K Actions or the app menu) is open, so `PalettePanel.sendEvent` swallows text-editing keystrokes the field editor would otherwise consume — the query must stay frozen while a menu owns the keyboard (matches Raycast). Untracked — read at event time, mirrored from the view's menu state.
+    @ObservationIgnored var menuOpen = false { didSet { onMenuOpenChanged?(menuOpen) } }
+    /// Fired when `menuOpen` flips so `PalettePanel` can hide/show the search field's caret while it keeps first-responder status (no focus swap, so the placeholder never reflows).
+    @ObservationIgnored var onMenuOpenChanged: ((Bool) -> Void)?
+
+    func prepare(mode: PaletteMode) {
+        self.mode = mode
+        query = ""
+        selection = 0
+        forceExpanded = false
+        hoverHighlightArmed = false
+        menuOpen = false
+        focusToken = UUID()
+        resetToken = UUID()
+    }
+}
