+++ Tinycast/Palette/PaletteWindowController.swift
