+++ Tinycast/Palette/RootPaletteView.swift
