+++ Tinycast/Windows/Dialog/DialogController.swift
