+++ Tinycast/Windows/Dialog/DialogPanel.swift
