+++ Tinycast/Windows/Dialog/DialogRequest.swift
