+++ Tinycast/Windows/Dialog/DialogView.swift
