+++ Tinycast/Windows/Dialog/VolumeSlider.swift
