+++ Tinycast/Windows/HUD/HUDPresenter.swift
