+++ Tinycast/Windows/HUD/MessageHUDController.swift
