+++ Tinycast/Windows/HUD/MessageHUDView.swift
